// src/synthetix/__generated__/graphql.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  BigDecimal: string
  BigInt: string
  Bytes: string
}
// ...
export type FuturesMarket = {
  __typename?: 'FuturesMarket'
  asset: Scalars['Bytes']
  id: Scalars['ID']
  isActive: Scalars['Boolean']
  marketKey: Scalars['Bytes']
  timestamp: Scalars['BigInt']
}
// ...
export type FuturesOrder = {
  __typename?: 'FuturesOrder'
  fee: Scalars['BigInt']
  futuresPosition?: Maybe<FuturesPosition>
  id: Scalars['ID']
  keeper: Scalars['Bytes']
  marginDelta: Scalars['BigInt']
  market: FuturesMarket
  orderId: Scalars['BigInt']
  orderType: FuturesOrderType
  size: Scalars['BigInt']
  status: FuturesOrderStatus
  targetPrice: Scalars['BigInt']
  targetRoundId: Scalars['BigInt']
  timestamp: Scalars['BigInt']
  trader: Trader
  txHash: Scalars['String']
}

export enum FuturesOrderStatus {
  Cancelled = 'Cancelled',
  Filled = 'Filled',
  Pending = 'Pending',
}

export enum FuturesOrderType {
  DelayedOffchainSubmitted = 'DelayedOffchainSubmitted',
  DelayedOrderSubmitted = 'DelayedOrderSubmitted',
  NextPriceOrderRemoved = 'NextPriceOrderRemoved',
  NextPriceOrderSubmitted = 'NextPriceOrderSubmitted',
}
// ...
export type FuturesPosition = {
  __typename?: 'FuturesPosition'
  avgEntryPrice: Scalars['BigInt']
  closeTimestamp?: Maybe<Scalars['BigInt']>
  entryPrice: Scalars['BigInt']
  exitPrice?: Maybe<Scalars['BigInt']>
  feesPaidToSynthetix: Scalars['BigInt']
  fundingIndex: Scalars['BigInt']
  id: Scalars['ID']
  initialMargin: Scalars['BigInt']
  isLiquidated: Scalars['Boolean']
  isOpen: Scalars['Boolean']
  lastPrice: Scalars['BigInt']
  leverage: Scalars['BigInt']
  long: Scalars['Boolean']
  margin: Scalars['BigInt']
  market: FuturesMarket
  netFunding: Scalars['BigInt']
  netTransfers: Scalars['BigInt']
  openTimestamp: Scalars['BigInt']
  realizedPnl: Scalars['BigInt']
  size: Scalars['BigInt']
  skew?: Maybe<Scalars['BigInt']>
  totalVolume: Scalars['BigInt']
  trader: Trader
  trades: Scalars['BigInt']
  txHash: Scalars['String']
  unrealizedPnl: Scalars['BigInt']
}
// ...
export type FuturesTrade = {
  __typename?: 'FuturesTrade'
  feesPaidToSynthetix: Scalars['BigInt']
  futuresOrder?: Maybe<FuturesOrder>
  futuresPosition: FuturesPosition
  id: Scalars['ID']
  margin: Scalars['BigInt']
  market: FuturesMarket
  netFunding: Scalars['BigInt']
  positionClosed: Scalars['Boolean']
  positionSize: Scalars['BigInt']
  price: Scalars['BigInt']
  realizedPnl: Scalars['BigInt']
  size: Scalars['BigInt']
  timestamp: Scalars['BigInt']
  trader: Trader
  txHash: Scalars['String']
  type: FuturesTradeType
}

export enum FuturesTradeType {
  Liquidated = 'Liquidated',
  PositionClosed = 'PositionClosed',
  PositionModified = 'PositionModified',
  PositionOpened = 'PositionOpened',
}
// ...
export type Trader = {
  __typename?: 'Trader'
  createdAt: Scalars['BigInt']
  feesPaidToSynthetix: Scalars['BigInt']
  id: Scalars['ID']
  lastTradeTimestamp?: Maybe<Scalars['BigInt']>
  margin: Scalars['BigInt']
  realizedPnl: Scalars['BigInt']
  totalLiquidations: Scalars['BigInt']
  totalMarginLiquidated: Scalars['BigInt']
  totalVolume: Scalars['BigInt']
  trades: Array<FuturesTrade>
}